Fix some issues in Remote Desktop cases
Updated OnAttached methods to make sure the UI can be scaled properly when application runs directly on different DPI of monitor(for example, using Surface 3 to remote control a 96 DPI PC and then run the program in remote desktop).

Updated UpdateDpiScaling methods to use another way to scale the UI, this fix issues which can not scale the UI properly when using a high DPI device(e.g. Surface 3) to remote control a low DPI PC which already running the program.